🔥 Pick the closest reasonable commit if the current one was not pushed
After this change there are now 3 actions:

- generating a link to the currently checked out commit
- generating a link to the latest commit from the remote's default branch
- generating a link to some parent commit that is definitely available at the remote

...with the last one becoming the default assigned to `Cmd`+`Shift`+`L`.

The main use cases are generating a link to whatever I see in my IDE (so,
close commits tend to also be acceptable since the code hasn't changed
much) or showing something that is currently in the latest master. Also, I
expect there to be not that many users relying on the plugin generating
a link to exactly the current commit in scenarios when it has not been
pushed, thought I don't really have any statistics. I assume no one has
tried automating anything by calling the action or the default shortcut.